+++ Client/Codes/TimerSystem.cpp
@@ -42,8 +42,8 @@ void TimerSystem::Update(const float& deltaTime)
     {
 
         float rate = (1 + int((timer - 30.f) / 5.f)) / 6.f;
-        int rates = 6 - ((timer - 30.f) / 5);
-        _RemainingUI[1]->SetOffsetPosition(Vector3(-10.8 * rates, 0.f, 0.f));
+        int rates = 6 - (int)((timer - 30.f) / 5);
+        _RemainingUI[1]->SetOffsetPosition(Vector3(-10.8f * rates, 0.f, 0.f));
         _RemainingUI[1]->SetScale({ 0.5f * rate, 1.0f, 0.f }); 
       //  _RemainingUI[1]->SetOffsetPosition(Vector3(-(126* 0.5f *0.5f* (1.f - rate)), 0.f, 0.f)); Èå¾Æ¾Æ¾Æ¾Ó
     }
@@ -51,8 +51,8 @@ void TimerSystem::Update(const float& deltaTime)
     {
         _RemainingUI[1]->SetScale({ 0.f,0.f,0.f });
         float rate = (1 + int(timer / 5.f)) / 6.f;
-        int rates = 6 - (timer / 5);
-        _RemainingUI[0]->SetOffsetPosition(Vector3(-10.8 * rates, 0.f, 0.f));
+        int rates = 6 - (int)(timer / 5);
+        _RemainingUI[0]->SetOffsetPosition(Vector3(-10.8f * rates, 0.f, 0.f));
         _RemainingUI[0]->SetScale({ 0.5f * rate, 1.0f, 0.f });
     }
